% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataset.R
\name{check_dataset}
\alias{check_dataset}
\title{checks and sanitizes a dataset}
\usage{
check_dataset(data)
}
\arguments{
\item{data}{data.frame containing data to be checked}
}
\value{
data.frame with checked and sanitized data
}
\description{
Checks if a given data contains correct values.

Deals with most common discrepancies and returns a sanitized dataset
which is ready to import.
}
