% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mismatch.R
\name{check_mismatch}
\alias{check_mismatch}
\title{checks if values fit in provided dict}
\usage{
check_mismatch(col, values, dict)
}
\arguments{
\item{col}{data column name}

\item{values}{vector of values}

\item{dict}{vector of allowed values}
}
\description{
Small helper for nicely reporting values outside of allowed set.
}
