% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_exec.R
\name{db_exec}
\alias{db_exec}
\title{simple helper for running SQL queries}
\usage{
db_exec(conn, query, param)
}
\arguments{
\item{conn}{DBI connection}

\item{query}{SQL query}

\item{param}{list of values to substitute}
}
\value{
data.frame with query results
}
\description{
simple helper for running SQL queries
}
