% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dataset.R
\name{import_dataset}
\alias{import_dataset}
\title{imports given dataset into the database}
\usage{
import_dataset(data, year, user, password)
}
\arguments{
\item{data}{data.frame returned by the \code{\link{check_dataset}}}

\item{year}{dataset year}

\item{user}{database user name}

\item{password}{database password}
}
\description{
Imports dataset prepared by the \code{\link{check_dataset}}
}
