% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dataset.R
\name{remove_dataset}
\alias{remove_dataset}
\title{removes given dataset from the database}
\usage{
remove_dataset(country, dataSource, user, password)
}
\arguments{
\item{country}{country name}

\item{dataSource}{data source name}

\item{user}{database user name}

\item{password}{database password}
}
\description{
removes given dataset from the database
}
